+++ R/class_builder.R
@@ -118,10 +118,20 @@ target_run.tar_builder <- function(target, envir, path_store) {
 }
 
 #' @export
-target_run_worker.tar_builder <- function(target, envir, path_store, options) {
-  tar_options$import(options)
+target_run_worker.tar_builder <- function(
+  target,
+  envir,
+  path_store,
+  fun,
+  options,
+  envvars
+) {
   envir <- if_any(identical(envir, "globalenv"), globalenv(), envir)
   tar_option_set(envir = envir)
+  tar_runtime$set_store(path_store)
+  tar_runtime$set_fun(fun)
+  tar_options$import(options)
+  set_envvars(envvars)
   target_gc(target)
   target_run(target, envir, path_store)
   builder_serialize_value(target)
@@ -431,13 +441,11 @@ builder_wait_correct_hash <- function(target) {
 builder_set_tar_runtime <- function(target, frames, path_store) {
   tar_runtime$set_target(target)
   tar_runtime$set_frames(frames)
-  tar_runtime$set_store(path_store)
 }
 
 builder_unset_tar_runtime <- function() {
   tar_runtime$unset_target()
   tar_runtime$unset_frames()
-  tar_runtime$unset_store()
 }
 
 builder_serialize_value <- function(target) {