+++ R/class_runtime.R
@@ -2,13 +2,15 @@ runtime_init <- function(
   target = NULL,
   frames = NULL,
   interactive = NULL,
-  store = NULL
+  store = NULL,
+  fun = NULL
 ) {
   runtime_new(
     target = target,
     frames = frames,
     interactive = interactive,
-    store = store
+    store = store,
+    fun = fun
   )
 }
 
@@ -16,13 +18,15 @@ runtime_new <- function(
   target = NULL,
   frames = NULL,
   interactive = NULL,
-  store = NULL
+  store = NULL,
+  fun = NULL
 ) {
   runtime_class$new(
     target = target,
     frames = frames,
     interactive = interactive,
-    store = store
+    store = store,
+    fun = fun
   )
 }
 
@@ -36,16 +40,19 @@ runtime_class <- R6::R6Class(
     frames = NULL,
     interactive = NULL,
     store = NULL,
+    fun = NULL,
     initialize = function(
       target = NULL,
       frames = NULL,
       interactive = NULL,
-      store = NULL
+      store = NULL,
+      fun = NULL
     ) {
       self$target <- target
       self$frames <- frames
       self$interactive <- interactive
       self$store <- store
+      self$fun <- fun
     },
     exists_target = function() {
       !is.null(self$target)
@@ -59,6 +66,9 @@ runtime_class <- R6::R6Class(
     exists_store = function() {
       !is.null(self$store)
     },
+    exists_fun = function() {
+      !is.null(self$fun)
+    },
     get_target = function() {
       self$target
     },
@@ -70,6 +80,9 @@ runtime_class <- R6::R6Class(
     },
     get_store = function() {
       self$store
+    },
+    get_fun = function() {
+      self$fun
     },
     set_target = function(target) {
       self$target <- target
@@ -83,6 +96,9 @@ runtime_class <- R6::R6Class(
     set_store = function(store) {
       self$store <- store
     },
+    set_fun = function(fun) {
+      self$fun <- fun
+    },
     unset_target = function() {
       self$target <- NULL
     },
@@ -94,6 +110,9 @@ runtime_class <- R6::R6Class(
     },
     unset_store = function() {
       self$store <- NULL
+    },
+    unset_fun = function() {
+      self$fun <- NULL
     },
     validate = function() {
       if (!is.null(self$target)) {
@@ -110,9 +129,29 @@ runtime_class <- R6::R6Class(
       if (!is.null(self$store)) {
         tar_assert_scalar(self$store)
         tar_assert_chr(self$store)
+        tar_assert_nzchar(self$store)
+      }
+      if (!is.null(self$fun)) {
+        tar_assert_scalar(self$fun)
+        tar_assert_chr(self$fun)
+        tar_assert_nzchar(self$fun)
       }
     }
   )
 )
+
+#' @title Get the `tar_runtime` object.
+#' @export
+#' @keywords internal
+#' @description For internal purposes only. Not a user-side function.
+#'   Do not invoke directly.
+#' @details Manages internal settings
+#'   that targets need while they run.
+#' @return The internal `tar_runtime` object of class `"tar_runtime"`.
+#' @examples
+#' tar_runtime_object()
+tar_runtime_object <- function() {
+  tar_runtime
+}
 
 tar_runtime <- runtime_init()