+++ R/tar_envvars.R
@@ -1,16 +1,27 @@
-#' @title Show environment variables to customize `targets`
+#' @title Show `targets` environment variables.
 #' @export
 #' @family configuration
-#' @description You can customize the behavior of `targets`
+#' @description Show all the special environment variables
+#'   available for customizing `targets`.
+#' @details You can customize the behavior of `targets`
 #'   with special environment variables. The sections in this help file
 #'   describe each environment variable, and the `tar_envvars()` function
 #'   lists their current values.
-#' @details If you modify environment variables, please set them
+#'
+#'   If you modify environment variables, please set them
 #'   in project-level `.Renviron` file so you do not lose your
 #'   configuration when you restart your R session.
 #'   Modify the project-level `.Renviron` file with
 #'   `usethis::edit_r_environ(scope = "project")`. Restart
 #'   your R session after you are done editing.
+#'
+#'   For targets that run on parallel workers
+#'   created by [tar_make_clustermq()] or [tar_make_future()],
+#'   only the environment variables listed by [tar_envvars()]
+#'   are specifically exported to the targets.
+#'   For all other environment variables, you will have to set
+#'   the values manually, e.g. a project-level `.Renviron` file
+#'   (for workers that have access to the local file system).
 #' @return A data frame with one row per environment variable
 #'   and columns with the name and current value of each.
 #'   An unset environment variable will have a value of `""`
@@ -54,6 +65,7 @@ tar_envvars <- function(unset = "") {
   names <- c(
     "TAR_ASK",
     "TAR_CONFIG",
+    "TAR_PROJECT",
     "TAR_WARN"
   )
   values <- map_chr(names, Sys.getenv, unset = unset, names = FALSE)