+++ R/tar_network.R
@@ -93,7 +93,9 @@ tar_network <- function(
     callr_function = callr_function,
     callr_arguments = callr_arguments,
     envir = envir,
-    script = script
+    script = script,
+    store = store,
+    fun = "tar_network"
   )
 }