+++ R/tar_prune.R
@@ -50,7 +50,9 @@ tar_prune <- function(
     callr_function = callr_function,
     callr_arguments = callr_arguments,
     envir = envir,
-    script = script
+    script = script,
+    store = store,
+    fun = "tar_prune"
   )
   invisible(out)
 }