+++ R/tar_sitrep.R
@@ -118,7 +118,9 @@ tar_sitrep <- function(
     callr_function = callr_function,
     callr_arguments = callr_arguments,
     envir = envir,
-    script = script
+    script = script,
+    store = store,
+    fun = "tar_sitrep"
   )
 }