+++ tests/testthat/test-tar_envvars.R
@@ -1,7 +1,11 @@
 tar_test("tar_envvars", {
   out <- tar_envvars()
   expect_equal(sort(colnames(out)), sort(c("name", "value")))
-  expect_true(all(c("TAR_ASK", "TAR_WARN") %in% out$name))
+  expect_true(
+    all(
+      c("TAR_ASK", "TAR_CONFIG", "TAR_PROJECT", "TAR_WARN") %in% out$name
+    )
+  )
   value <- Sys.getenv("TAR_ASK")
   on.exit(Sys.setenv(TAR_ASK = value))
   Sys.unsetenv("TAR_ASK")