+++ tests/testthat/test-tar_make.R
@@ -165,6 +165,30 @@ tar_test("custom script and store args with callr function", {
   expect_true(file.exists("_targets.yaml"))
 })
 
+tar_test("runtime settings are forwarded, local process", {
+  skip_on_cran()
+  tar_script({
+    writeLines(tar_store(), "store.txt")
+    tar_target(x, 1)
+  })
+  tar_make(callr_function = NULL, store = "custom_store")
+  expect_equal(readLines("store.txt"), "custom_store")
+})
+
+tar_test("runtime settings are forwarded, extrernal process", {
+  skip_on_cran()
+  tar_script({
+    writeLines(tar_store(), "store.txt")
+    tar_target(x, 1)
+  })
+  tar_make(
+    store = "custom_store",
+    reporter = "silent",
+    callr_arguments = list(spinner = FALSE)
+  )
+  expect_equal(readLines("store.txt"), "custom_store")
+})
+
 tar_test("null environment", {
   tar_script(tar_target(x, "x"))
   tar_make(callr_function = NULL, envir = NULL)